pyproject.toml
[project]
name = "agent-workspace"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pulumi>=3.0.0,<4.0.0",
    "pulumi-gcp>=9.0.0,<10.0.0",
    "python-dotenv>=1.2.1",
]
